decomp: C batch (+2 functions, 402/633 in src/matched)
- sub_0806B064 (old_agbcc): tail assign keeps total live -> r3/r4 split
- sub_08062CF4 (old_agbcc): hoisted palette base pointer -> r5/r6 split
- park improved sub_080604C8 seed (92/112, same size)
- include sweep_compilers.py / asm_c.py helpers

+++ include/data_symbols.h
@@ -48,7 +48,7 @@ extern u8 gData_080B738E[];
 extern u32 gData_03000198[];
 extern u8 gData_03000290[];
 extern u32 gData_030003CC[];
-extern u8 gData_05000200[];
+extern u16 gData_05000200[];
 extern u32 gData_08079068[];
 extern u32 gData_08079358[];
 extern u8 gData_030002A0[];
